latex-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
supabase==2.0.3
python-multipart==0.0.6
pillow===11.0.0
aiohttp===3.10.10
aiofiles===24.1.0